Remove wai-bindgen-wasmer crate
This commit removes the wai-bindgen-wasmer crate, which was migrated to
the https://github.com/wasmerio/wasmer repository.

This move was forced by Wasmer adopting wai for internal use, which made
not having the bindings in the same repo unfeasible.

In the medium term we want to rewrite wai-bindgen-gen-wasmer to make the crate redundant.

Note that the tests now rely on the published version of wai-bindgen-wasmer.

Closes #31